temp/sandbox/requirements.txt
# API et AI
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
anthropic==0.39.0
openai==1.54.0

# Web scraping et veille
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
httpx==0.26.0

# Data processing
pandas==2.2.0
python-dotenv==1.0.1

# LinkedIn API (unofficial)
linkedin-api==2.2.0

# RSS et newsletters
email-reply-parser==0.5.12
html2text==2024.2.26

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
starlette==0.38.2
slowapi==0.1.9

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
itsdangerous==2.1.2

# Utilities
pydantic>=2.10.0
tenacity==8.2.3
jinja2==3.1.3
python-dateutil==2.8.2
pytz==2024.1
python-pptx==1.0.2
PyPDF2==3.0.1
google-generativeai>=0.8.0
Pillow>=10.2.0
